pyproject.toml
[project]
name = "traffic-ai"
version = "0.1.0"
description = "Hardware-adaptive intelligent transportation system"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg",
    "alembic",
    "celery[redis]>=5.3",
    "redis>=5.0",
    "flower",
    "influxdb-client[async]>=1.40",
    "aiohttp>=3.9",
    "httpx>=0.27",
    "psutil",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "pydantic>=2.5",
    "pydantic-settings",
    "geoalchemy2>=0.14",
    "shapely>=2.0",
    "numpy",
    "pandas",
    "onnxruntime",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "aiosqlite",
]
ml = [
    "shap",
    "ultralytics",
    "boto3",
]
s3 = [
    "boto3",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
